Introduce new modules to handle nav click events and book management
