+++ index.html
@@ -52,6 +52,6 @@
     <footer>
       <span>Copyright</span>
     </footer>
-    <script src="index.js"></script>
+    <script type="module" src="index.js"></script>
   </body>
 </html>
